Refactor CounterList and App
