+++ src/app/containers/App.jsx
@@ -1,18 +1,31 @@
 import React, { Component } from 'react';
-import Counter from './Counter';
-import { createStore, combineReducers } from 'redux';
+import CounterApp from './CounterApp';
+import { createStore, applyMiddleware, combineReducers, compose } from 'redux';
+import { devTools, persistState } from 'redux-devtools';
+import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
+import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import * as reducers from '../reducers';
 
+const finalCreateStore = compose(
+  applyMiddleware(thunk)
+)(createStore);
+
 const reducer = combineReducers(reducers);
-const store = createStore(reducer);
+const store = finalCreateStore(reducer);
+
+if (module.hot) {
+  module.hot.accept('../reducers', () =>
+    store.replaceReducer(combineReducers(require('../reducers')))
+  );
+}
 
 export default class App extends Component {
 
   render() {
      return (
        <Provider store={store}>
-         {() => <Counter /> }
+         {() => <CounterApp /> }
        </Provider>
      );
    }
